+++ ENTITIES/DTOs/HistoryDto.cs
@@ -8,12 +8,12 @@ namespace ENTITIES.DTOs
 {
     public class HistoryDto
     {
-        public ulong Id { get; set; }
+        public long Id { get; set; }
         public int LastChapter { get; set; }
         public int LastPage { get; set; }
         public DateTime UpdatedAt { get; set; }
-        public int UserId { get; set; }
-        public int MangaId { get; set; }
+        public long UserId { get; set; }
+        public long MangaId { get; set; }
         
         // Optional: Include user and manga info for display purposes
         public string? UserNickname { get; set; }
