+++ ENTITIES/DTOs/ReviewDto.cs
@@ -9,14 +9,14 @@ namespace ENTITIES.DTOs
 {
     public class ReviewDto
     {
-        public ulong Id { get; set; }
+        public long Id { get; set; }
         
         [Range(1, 5)]
         public int Stars { get; set; }
         public string? Text { get; set; }
         public DateTime CreationTime { get; set; }
-        public int UserId { get; set; }
-        public int MangaId { get; set; }
+        public long UserId { get; set; }
+        public long MangaId { get; set; }
         
         // Optional: Include user and manga info for display purposes
         public string? UserNickname { get; set; }
