+++ ENTITIES/Models/MangaGenre.cs
@@ -8,13 +8,13 @@ namespace ENTITIES.Models
 {
     public class MangaGenre
     {
-        public int Id { get; set; }  // PK
+        public long Id { get; set; }  // PK
         public int Order { get; set; } //not null
 
 
         //FK and References
-        public int MangaId { get; set; } //not null
-        public int GenreId { get; set; } //not null
+        public long MangaId { get; set; } //not null
+        public long GenreId { get; set; } //not null
 
         public Manga Manga { get; set; } = null!;
         public Genre Genre { get; set; } = null!;
